requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
boto3==1.28.44
Werkzeug==3.0.1
psycopg2-binary==2.9.9
python-dotenv==1.0.0
Pillow==10.0.1
geoalchemy2
shapely
